requirements.txt
address==0.1.1
amqp==2.5.2
antiorm==1.2.1
anyjson==0.3.3
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.1.0
black==19.10b0
celery==4.4.0
Click==7.0
db==0.1.1
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-rest-auth==0.9.5
django-service-objects==0.5.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
gunicorn==20.0.4
importlib-metadata==1.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
MyApplication==0.1.0
pathspec==0.6.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
regex==2019.12.9
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
vine==1.3.0
Werkzeug==0.16.0
whitenoise==5.0.1
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
Django==2.2.7
django-rest-auth==0.9.5
django-service-objects==0.5.0
djangorestframework==3.11.0
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
pkg-resources==0.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
